+++ src/Event.cpp
@@ -10,19 +10,19 @@
 
 using namespace AmpGen; 
 
-Event::Event( const size_t& N, const size_t& cacheSize) : 
+Event::Event( const unsigned& N, const unsigned& cacheSize) : 
   m_event(N), 
   m_cache(cacheSize) {
   }
 
-Event::Event( const real_t* data, const size_t& N, const size_t& cacheSize) :
+Event::Event( const real_t* data, const unsigned& N, const unsigned& cacheSize) :
   m_event(data, data+N),
   m_cache(cacheSize) {
   }
 
 void Event::print() const {
-  size_t nParticles = m_event.size()/4;
-  for( size_t i = 0 ; i < nParticles; ++i ){
+  unsigned nParticles = m_event.size()/4;
+  for( unsigned i = 0 ; i < nParticles; ++i ){
     double px = m_event[4*i+0];
     double py = m_event[4*i+1];
     double pz = m_event[4*i+2];
@@ -30,25 +30,25 @@ void Event::print() const {
     double s = pE*pE - px*px -py*py -pz*pz;
     INFO( "P["<<i<<"] = {"<<px<<", " << py << ", " << pz << ", " << pE << "}, m = " << sqrt(fabs(s)) );
   }
-  for( size_t i = 4*nParticles; i != m_event.size(); ++i ){
+  for( unsigned i = 4*nParticles; i != m_event.size(); ++i ){
     INFO( m_event[i] );
   }
 }
-real_t Event::s( const size_t& index ) const {
+real_t Event::s( const unsigned& index ) const {
   return get(4*index+3) * get(4*index+3)
     - get(4*index+0) * get(4*index+0)
     - get(4*index+1) * get(4*index+1)
     - get(4*index+2) * get(4*index+2);
 }
 
-real_t Event::s( const size_t& index1, const size_t& index2 ) const {
+real_t Event::s( const unsigned& index1, const unsigned& index2 ) const {
   return ( get(4*index1+3) + get(4*index2+3) )*( get(4*index1+3) + get(4*index2+3) ) - 
     ( get(4*index1+0) + get(4*index2+0) )*( get(4*index1+0) + get(4*index2+0) ) -
     ( get(4*index1+1) + get(4*index2+1) )*( get(4*index1+1) + get(4*index2+1) ) -
     ( get(4*index1+2) + get(4*index2+2) )*( get(4*index1+2) + get(4*index2+2) ) ;
 }
 
-real_t Event::s( const size_t& index1, const size_t& index2, const size_t& index3 ) const {
+real_t Event::s( const unsigned& index1, const unsigned& index2, const unsigned& index3 ) const {
   return ( get(4*index1+3) + get(4*index2+3) + get(4*index3+3) ) *
     ( get(4*index1+3) + get(4*index2+3) + get(4*index3+3) ) - 
     ( get(4*index1+0) + get(4*index2+0) + get(4*index3+0) ) *
@@ -59,7 +59,7 @@ real_t Event::s( const size_t& index1, const size_t& index2, const size_t& index
     ( get(4*index1+2) + get(4*index2+2) + get(4*index3+2) ) ;
 }
 
-real_t Event::s( const std::vector<size_t>& indices ) const {
+real_t Event::s( const std::vector<unsigned>& indices ) const {
   if( indices.size() == 2 ) return s( indices[0], indices[1] );
   if( indices.size() == 3 ) return s( indices[0], indices[1], indices[2] );
   double E=0;        double px=0;
@@ -79,18 +79,18 @@ void Event::printCache() const {
   }
 }
 
-void Event::set( const size_t& i, const std::vector<real_t>& p ){
-  for( size_t j = 0 ; j < 4; ++j) m_event[4*i + j ] = p[j];
+void Event::set( const unsigned& i, const std::vector<real_t>& p ){
+  for( unsigned j = 0 ; j < 4; ++j) m_event[4*i + j ] = p[j];
 }
-void Event::set( const size_t& i, const real_t* p ){
-  for( size_t j = 0 ; j < 4; ++j) m_event[4*i + j ] = p[j];
+void Event::set( const unsigned& i, const real_t* p ){
+  for( unsigned j = 0 ; j < 4; ++j) m_event[4*i + j ] = p[j];
 }
 void Event::set( const real_t* evt ){
-  for( size_t i = 0 ; i < m_event.size(); ++i ) m_event[i] = *(evt + i );
+  for( unsigned i = 0 ; i < m_event.size(); ++i ) m_event[i] = *(evt + i );
 }
-void Event::set( const size_t& i, const real_t& p ){ m_event[i] = p ; } 
+void Event::set( const unsigned& i, const real_t& p ){ m_event[i] = p ; } 
 
-void Event::swap( const unsigned int& i , const unsigned int& j )
+void Event::swap( const unsigned& i , const unsigned& j )
 {
   double tmp[4];
   std::memmove( tmp, &m_event[4*j], sizeof(tmp)); 
@@ -98,8 +98,8 @@ void Event::swap( const unsigned int& i , const unsigned int& j )
   std::memmove( &m_event[4*i], &tmp,sizeof(tmp));
 }
 
-void Event::setCache(const complex_t& value, const size_t& pos){ m_cache[pos] = value; }
-void Event::setCache( const std::vector<complex_t>& value, const size_t& pos )
+void Event::setCache(const complex_t& value, const unsigned& pos){ m_cache[pos] = value; }
+void Event::setCache( const std::vector<complex_t>& value, const unsigned& pos )
 {
   std::memmove( m_cache.data() + pos, value.data(), sizeof(complex_t) * value.size() );
 }
